requirements.txt
together
python-dotenv
pydantic
pydantic-settings
tqdm